fix(CLI args): found issue related to number of CLI args and the Test runWeather
If only one file is called, the other causes a NullPointer.
The fix is quiet easy by only working with one file at the same time.
This is a rather simple fix. However, multi-file support could easily
be added by allowing pairwise arguments (type and filepath) and
repeating the reading and output in a loop.

This is, however, not the most simple fix, since the issue could have
been work around by adding if statements to the outputs in main.